Fix parent/child warnings
The parent child warnings described in #83 were the result of us
trying to stuff fileNode into Mdx graphql types. As it turns out, this
is not necessary because you can query the parent using fragments
which more accurately reflect the existence or non-existence of parent
types. This also means you could query the base contentful/remote CMS
types when trying to add slug fields, etc.

Additionally, we unified the code paths that create nodes for all
sources and transformers. Since we now use the public APIs, we can
have more confidence that people can do everything we can in core when
creating new content types.